# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Kelompok API
  summary: Public and org-admin endpoints for the Kelompok alpha release.
  description: |
    Kelompok is an open-source organization CRM. This document is the contract
    for the implemented public read endpoints, the auth endpoints, and the
    `/api/v1/org-admin` endpoints that ship in the current alpha release.

    The contract intentionally tracks only the routes that are wired up in
    `internal/httpapi/server.go`. Endpoints that appear in `docs/04-api-cli.md`
    but are not yet implemented (events, donation campaigns, claim email and
    instagram verification, etc.) are out of scope until the corresponding
    handler ships.

    Response envelope conventions:

    * Success responses always include `data` and `message`. List endpoints
      also include a `meta` object with `count` and `limit`.
    * Error responses always carry an `error` object with `code`, `message`,
      and an optional `details` payload.

    Auth on `/api/v1/org-admin/*` accepts either a user session bearer token
    from `POST /api/v1/auth/login` or the operations fallback admin key sent
    as `X-Kelompok-Admin-Key`. When `KELOMPOK_ADMIN_ORGANIZATION_SLUGS` is
    configured, scoped admin keys must hit organization-scoped routes; global
    list routes return `403 admin_org_scope_required`.
  version: 1.0.0-alpha.2
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  contact:
    name: Kelompok maintainers
...
servers:
...
security: []
paths:
...
  /api/v1/organizations:
    get:
      tags: [Public — Organizations]
      summary: List public organizations
      operationId: listPublicOrganizations
      parameters:
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: A page of public organizations.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicOrganizationListResponse"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/organizations/{slug}:
    get:
      tags: [Public — Organizations]
      summary: Get a public organization by slug
      operationId: getPublicOrganization
      parameters:
        - $ref: "#/components/parameters/OrganizationSlug"
      responses:
        "200":
          description: Public organization profile, including active relationships.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicOrganizationDetailResponse"
        "404":
          $ref: "#/components/responses/OrganizationNotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/organizations/{slug}/claims:
    post:
      tags: [Public — Organizations]
      summary: Submit an organization claim request
      description: |
        Public-side intake for organization claims. The request creates (or
        reuses) a viewer-level user keyed off `requester_email`, then records
        a `pending` claim that an org-admin reviewer can later approve or
        reject via `/api/v1/org-admin/claims/{id}/{approve|reject}`.
      operationId: createOrganizationClaim
      parameters:
        - $ref: "#/components/parameters/OrganizationSlug"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClaimInput"
      responses:
        "201":
          description: Claim request created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClaimRequestResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/OrganizationNotFound"
  /api/v1/organization-onboarding-requests:
    post:
      tags: [Public — Organizations]
      summary: Submit a new organization onboarding request
      description: |
        User-authenticated intake for organizations that do not yet have a
        public profile. The server creates an organization with
        `claim_status=pending` plus a linked pending claim request for the
        caller. Owner access is not granted until an org-admin reviewer
        approves the claim through the existing review endpoint.
      operationId: createOrganizationOnboardingRequest
      security:
        - sessionBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrganizationOnboardingRequestInput"
      responses:
        "201":
          description: Pending organization and claim request created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationOnboardingRequestResponse"
        "400":
          description: |
            Invalid payload. Stable error codes include
            `organization_name_required`,
            `organization_official_email_invalid`,
            `organization_json_invalid`,
            `organization_onboarding_claim_method_invalid`,
            `organization_onboarding_claim_target_required`, and
            `organization_onboarding_claim_target_invalid`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/SessionRequired"
        "409":
          description: Organization slug already exists or is pending review.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/organizations/{slug}/posts:
    get:
      tags: [Public — Organizations]
      summary: List published posts for an organization
      operationId: listOrganizationPosts
      parameters:
        - $ref: "#/components/parameters/OrganizationSlug"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: A page of published posts for this organization.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicPostListResponse"
        "404":
          $ref: "#/components/responses/OrganizationNotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/organizations/{slug}/posts/{post_slug}:
    get:
      tags: [Public — Organizations]
      summary: Get a published post scoped to an organization
      operationId: getOrganizationPost
      parameters:
        - $ref: "#/components/parameters/OrganizationSlug"
        - $ref: "#/components/parameters/PostSlug"
      responses:
        "200":
          description: Published post for this organization.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicPostDetailResponse"
        "404":
          description: Organization or post not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/organizations/{slug}/impact-reports:
    get:
      tags: [Public — Impact Reports]
      summary: List published impact reports for an organization
      operationId: listOrganizationImpactReports
      parameters:
        - $ref: "#/components/parameters/OrganizationSlug"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: A page of published impact reports for this organization.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicImpactReportListResponse"
        "404":
          $ref: "#/components/responses/OrganizationNotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/posts:
    get:
      tags: [Public — Posts]
      summary: List published posts
      operationId: listPublicPosts
      parameters:
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: A page of published posts.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicPostListResponse"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/posts/{slug}:
    get:
      tags: [Public — Posts]
      summary: Get a published post by slug
      description: |
        When a slug is shared by more than one organization, the API returns
        `409 post_slug_ambiguous` and the caller should use the
        organization-scoped post detail endpoint instead.
      operationId: getPublicPost
      parameters:
        - $ref: "#/components/parameters/PostSlug"
      responses:
        "200":
          description: Published post.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicPostDetailResponse"
        "404":
          description: Post not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Post slug is shared by more than one organization.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      description: |
        Creates a viewer-level user. If a row already exists with this email
        and no password set yet (e.g. created during a public claim intake),
        the password is set on that record.
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User created or upgraded with a password.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: User already has a password set.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Log in and obtain a session token
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Session created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          description: Invalid email or password.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/auth/google:
    post:
      tags: [Auth]
      summary: Log in or register with Google OAuth code
      operationId: googleLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoogleLoginInput"
      responses:
        "200":
          description: Session created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/auth/logout:
    post:
      tags: [Auth]
      summary: Revoke the current session
      operationId: logoutUser
      security:
        - sessionBearer: []
      responses:
        "200":
          description: Session revoked.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogoutResponse"
        "401":
          $ref: "#/components/responses/SessionRequired"
  /api/v1/auth/me:
    get:
      tags: [Auth]
      summary: Get the current user and organization roles
      operationId: getMe
      security:
        - sessionBearer: []
      responses:
        "200":
          description: Current user and the organization roles attached to it.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
        "401":
          $ref: "#/components/responses/SessionRequired"
        "500":
          $ref: "#/components/responses/InternalError"
    patch:
      tags: [Auth]
      summary: Update the current user's profile
      operationId: updateMe
      security:
        - sessionBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileInput"
      responses:
        "200":
          description: Updated user and current organization roles.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/SessionRequired"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/org-admin/organizations:
    get:
      tags: [Org Admin — Organizations]
      summary: List organizations (admin)
      description: |
        Global list. Scoped admin keys (when
        `KELOMPOK_ADMIN_ORGANIZATION_SLUGS` is configured) and non-superadmin
        user sessions must use the organization-scoped routes instead.
      operationId: listAdminOrganizations
      security:
        - sessionBearer: []
        - adminApiKey: []
      parameters:
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: A page of admin organization records.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOrganizationListResponse"
        "401":
          $ref: "#/components/responses/AdminAuthRequired"
        "403":
          $ref: "#/components/responses/AdminOrgScopeRequired"
        "503":
          $ref: "#/components/responses/AdminAuthNotConfigured"
    post:
      tags: [Org Admin — Organizations]
      summary: Create an organization (admin)
      description: |
        Creates a new organization profile through the admin contract. Public
        users cannot create organizations directly through this endpoint; they
        must use claim/onboarding flows exposed by the product UI. `name` is
        required, `slug` is normalized from the supplied slug or generated from
        `name`, and `claim_status` defaults to `unclaimed`. Duplicate slugs
        return `409 organization_slug_taken`; invalid names, claim status,
        official email, or JSON object fields return stable 400 error codes.
      operationId: createAdminOrganization
      security:
        - sessionBearer: []
        - adminApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminOrganizationInput"
      responses:
        "201":
          description: Organization created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOrganizationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/AdminAuthRequired"
        "403":
          $ref: "#/components/responses/AdminOrgForbidden"
        "409":
          $ref: "#/components/responses/OrganizationSlugTaken"
  /api/v1/org-admin/organizations/{slug}:
    parameters:
      - $ref: "#/components/parameters/OrganizationSlug"
    get:
      tags: [Org Admin — Organizations]
      summary: Get an organization by slug (admin)
      operationId: getAdminOrganization
      security:
        - sessionBearer: []
        - adminApiKey: []
      responses:
        "200":
          description: Admin organization record.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOrganizationResponse"
        "401":
          $ref: "#/components/responses/AdminAuthRequired"
        "403":
          $ref: "#/components/responses/AdminOrgForbidden"
        "404":
          $ref: "#/components/responses/OrganizationNotFound"
    patch:
      tags: [Org Admin — Organizations]
      summary: Update an organization (admin)
      operationId: updateAdminOrganization
      security:
        - sessionBearer: []
        - adminApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminOrganizationInput"
      responses:
        "200":
          description: Updated organization.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOrganizationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/AdminAuthRequired"
        "403":
          $ref: "#/components/responses/AdminOrgForbidden"
        "404":
          $ref: "#/components/responses/OrganizationNotFound"
  /api/v1/org-admin/organizations/{slug}/relationships:
    get:
      tags: [Org Admin — Relationships]
      summary: List relationships for an organization (admin)
      operationId: listOrganizationRelationships
      security:
        - sessionBearer: []
        - adminApiKey: []
      parameters:
        - $ref: "#/components/parameters/OrganizationSlug"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Relationships where this organization is parent or child.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RelationshipListResponse"
        "401":
          $ref: "#/components/responses/AdminAuthRequired"
        "403":
          $ref: "#/components/responses/AdminOrgForbidden"
        "404":
          $ref: "#/components/responses/OrganizationNotFound"
...
components:
  securitySchemes:
    sessionBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        User session token returned by `POST /api/v1/auth/login`. Send as
        `Authorization: Bearer <token>`.
    adminApiKey:
      type: apiKey
      in: header
      name: X-Kelompok-Admin-Key
      description: |
        Operations fallback admin key (`KELOMPOK_ADMIN_API_KEY`). When
        `KELOMPOK_ADMIN_ORGANIZATION_SLUGS` is set, the key is restricted to
        those slugs and cannot use global list routes without an
        `organization_slug` filter.
...
  responses:
...
    InternalError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...